+++ Common-Core/0-Libft/ft_strnstr.c
@@ -6,16 +6,16 @@
 /*   By: bpolat <bpolat@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/11 14:21:52 by bpolat            #+#    #+#             */
-/*   Updated: 2024/10/11 14:23:48 by bpolat           ###   ########.fr       */
+/*   Updated: 2024/10/18 13:23:47 by bpolat           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_strnstr(const char *hs, const char *nd, int len)
+char	*ft_strnstr(const char *hs, const char *nd, size_t len)
 {
-	int	i;
-	int	j;
+	size_t	i;
+	int		j;
 
 	if (nd[0] == '\0')
 		return ((char *)hs);
